requirements.txt
fastapi
uvicorn 
jinja2 
mysql-connector-python 
python-multipart 
whisper 
SpeechRecognition 
groq
pandas
pyjwt
python-docx
#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